solutions report: moves blm calibrations to another page
